fix searchbar sync on nav w/controlled component
